// src/api/generated/model/addressResponse.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Ecommerce Copa API
 * API backend do Ecommerce Copa com autenticação por sessão, catálogo público e CRUD administrativo.
 * OpenAPI spec version: 1.0.0
 */

export interface AddressResponse {
	id?: string;
	name?: string;
	street?: string;
	number?: string;
	state?: string;
	city?: string;
	neighborhood?: string;
	complement?: string;
	postalCode?: string;
	isDefault?: boolean;
}