// src/api/generated/model/createAddressRequest.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Ecommerce Copa API
 * API backend do Ecommerce Copa com autenticação por sessão, catálogo público e CRUD administrativo.
 * OpenAPI spec version: 1.0.0
 */

export interface CreateAddressRequest {
	/**
	 * @minLength 0
	 * @maxLength 80
	 */
	name: string;
	/**
	 * @minLength 0
	 * @maxLength 160
	 */
	street: string;
	/**
	 * @minLength 0
	 * @maxLength 30
	 */
	number: string;
	/**
	 * @minLength 0
	 * @maxLength 80
	 */
	state: string;
	/**
	 * @minLength 0
	 * @maxLength 120
	 */
	city: string;
	/**
	 * @minLength 0
	 * @maxLength 120
	 */
	neighborhood: string;
	/**
	 * @minLength 0
	 * @maxLength 255
	 */
	complement?: string;
	/**
	 * @minLength 0
	 * @maxLength 20
	 */
	postalCode: string;
	isDefault?: boolean;
}